requirements.txt
torch>=1.12.0
numpy>=1.21.0
Pillow>=9.0.0
tqdm>=4.60.0
scipy>=1.7.0
einops>=0.6.0
omegaconf>=2.1.0
huggingface_hub>=0.14.0
